+++ pkg/reconciler/events/doc.go
@@ -1,0 +1,21 @@
+/*
+Copyright 2020 The Tekton Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// Package events defines the interface and different implementations for events
+//
+// Tekton uses events to notify about changes in Condition of resources, like a
+// TaskRun is starting, a PipelineRun failed to run or could not be validated.
+package events
